integer for all tables and shipping order

+++ database/migrations/2020_06_13_061111_create_products_table.php
@@ -18,8 +18,8 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->string('slug');
             $table->text('description');
-            $table->decimal('price');
-            $table->decimal('weight');
+            $table->integer('price');
+            $table->integer('weight');
             $table->integer('quantity');
             $table->integer('quantity_alert')->default(5);
             $table->boolean('active')->default(false);
