+++ database/migrations/2020_06_14_062625_create_ranges_table.php
@@ -15,7 +15,7 @@ class CreateRangesTable extends Migration
     {
         Schema::create('ranges', function (Blueprint $table) {
             $table->id();
-            $table->decimal('max');
+            $table->integer('max');
         });
     }
